refactor: Extract util for finding all classes used in openapi.json
Removes the Aot processor that finds endpoint classes as they are also detected from openapi.json
