+++ packages/java/endpoint/src/main/resources/META-INF/spring/aot.factories
@@ -1,2 +1,1 @@
-org.springframework.beans.factory.aot.BeanFactoryInitializationAotProcessor=dev.hilla.EndpointInitializationAotProcessor
 org.springframework.aot.hint.RuntimeHintsRegistrar=dev.hilla.springnative.HillaHintsRegistrar
