+++ packages/java/endpoint/src/test/resources/dev/hilla/openapi-nocustomtypes.json
