+++ packages/java/endpoint/src/test/resources/dev/hilla/openapi-referring-customtypes.json
